+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - text can be dragged into table to create a new snippet
 - snippets can be dragged directly into editor and the code are copied
 - added "save as file" feature
+- added "clone snippet" feature
 
 ### Fixed
 - show a dialog error if verification code is not correct